requirements.txt
Flask
google-genai
python-dotenv
gunicorn
flask-cors